+++ test_support/HeaderHygiene.cpp
@@ -27,6 +27,7 @@
 #include <felitronics/convolution/PartitionedConvolver.h>
 #include <felitronics/core/Config.h>
 #include <felitronics/core/DelayLine.h>
+#include <felitronics/core/DeliveryResampler.h>
 #include <felitronics/core/DryAligner.h>
 #include <felitronics/core/Fft.h>
 #include <felitronics/core/FlushToZero.h>